Intercambio+cambios seguridad Jorge

+++ HiberLibros/src/main/java/com/hiberlibros/HiberLibros/services/UsuarioService.java
@@ -1,12 +1,14 @@
-
 package com.hiberlibros.HiberLibros.services;
 
 import com.hiberlibros.HiberLibros.entities.Usuario;
+import com.hiberlibros.HiberLibros.interfaces.ISeguridadService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.hiberlibros.HiberLibros.repositories.UsuarioRepository;
 import com.hiberlibros.HiberLibros.interfaces.IUsuarioService;
+import java.util.Optional;
+import javax.transaction.Transactional;
 
 /**
  *
@@ -17,31 +19,47 @@ public class UsuarioService implements IUsuarioService{
     @Autowired
     private UsuarioRepository urService;
 
+    @Autowired
+    private ISeguridadService serviceUsuarioSeguridad;
+
+    @Override
+    public String guardarUsuarioYSeguridad(Usuario u, String password) {
+        String resultado = guardarUsuario(u);
+        Optional<Usuario> usu = urService.findByMail(u.getMail());
+        if (usu.isPresent()) {
+            serviceUsuarioSeguridad.altaUsuarioSeguridad(u.getMail(), u.getId(), password, "Usuario");
+        }
+
+        return resultado;
+    }
+
     @Override
     public String guardarUsuario(Usuario u) {
-        String resultado="";
-        int auxMail=u.getMail().indexOf("@");
-        String mailSubstring=u.getMail().substring(auxMail);//para comprar si el mail tiene buen formato
-        if(u.getNombre()==null || u.getApellido()==null || u.getDireccion()==null  || u.getCiudad()==null  || u.getMail()==null || u.getTelef()==null){
-            resultado="Error: Campo requerido vacío";
-        }
-        else if(!mailSubstring.contains(".")){ //@blabla. si no punto consideramos que no esta bien
-            resultado="Error: e-mail incorrecto";
-        }
-        else if(urService.findByMailContainsIgnoreCase(u.getMail()).isPresent()){//su ya existe ese mail
-            resultado="Error: Ya existe un usuario registrado con ese e-mail";
-        }
-        else{
+        String resultado = "";
+        int auxMail = u.getMail().indexOf("@");
+        String mailSubstring = u.getMail().substring(auxMail);//para comprar si el mail tiene buen formato
+        if (u.getNombre() == null || u.getApellido() == null || u.getDireccion() == null || u.getCiudad() == null || u.getMail() == null || u.getTelef() == null) {
+            resultado = "Error: Campo requerido vacío";
+        } else if (!mailSubstring.contains(".")) { //@blabla. si no punto consideramos que no esta bien
+            resultado = "Error: e-mail incorrecto";
+        } else if (urService.findByMailContainsIgnoreCase(u.getMail()).isPresent()) {//su ya existe ese mail
+            resultado = "Error: Ya existe un usuario registrado con ese e-mail";
+        } else {
             urService.save(u);
-            resultado="Usuario registrado con éxito";
+            resultado = "Usuario registrado con éxito";
         }
         return resultado;
 
     }
 
     @Override
+    @Transactional
     public void borrarUsuario(Integer id) {
-       urService.deleteById(id);
+        Optional<Usuario> usuario = urService.findById(id);
+        if (usuario.isPresent()) {
+            urService.deleteById(id);
+            serviceUsuarioSeguridad.bajaUsuarioSeguridadPorMail(usuario.get().getMail());
+        }
     }
 
     @Override
@@ -51,10 +69,9 @@ public class UsuarioService implements IUsuarioService{
 
     @Override
     public boolean registrado(String mail) { //comprueba si existe ese usuario por mail
-        if(urService.findByMail(mail).isEmpty()){
+        if (urService.findByMail(mail).isEmpty()) {
             return false;
-        }
-        else{
+        } else {
             return true;
         }
     }
@@ -74,5 +91,5 @@ public class UsuarioService implements IUsuarioService{
     public Usuario usuarioId(Integer id) {
         return urService.findById(id).get();
     }
-    
+
 }
